finishing up logic, adding functional test

+++ app/src/main/java/com/yminachi/pathofleastresistance/leastresistancepathcalculator/GridConverter.java
@@ -3,27 +3,27 @@ package com.yminachi.pathofleastresistance.leastresistancepathcalculator;
 import org.apache.commons.math.linear.RealMatrix;
 import org.apache.commons.math.linear.RealMatrixImpl;
 
-import static java.lang.Integer.*;
+import static java.lang.Double.parseDouble;
 
 public class GridConverter {
-    public RealMatrix convertToGrid(String input){
+    public RealMatrix convertToGrid(String input) {
         String[] rows = input.split("\n");
 
         double[][] grid = new double[rows.length][];
 
-        for(int rowNumber = 0; rowNumber < rows.length; rowNumber++){
+        for (int rowNumber = 0; rowNumber < rows.length; rowNumber++) {
             grid[rowNumber] = convertRowToArray(rows[rowNumber]);
         }
 
         return new RealMatrixImpl(grid);
     }
 
-    private double[] convertRowToArray(String row){
+    private double[] convertRowToArray(String row) {
         String[] cellValues = row.split(" ");
         double[] rowCellValues = new double[cellValues.length];
 
-        for(int columnIndex = 0; columnIndex < cellValues.length; columnIndex++){
-            rowCellValues[columnIndex] = parseInt(cellValues[columnIndex]);
+        for (int columnIndex = 0; columnIndex < cellValues.length; columnIndex++) {
+            rowCellValues[columnIndex] = parseDouble(cellValues[columnIndex]);
         }
 
         return rowCellValues;
